Create mighty stand
Created a stand for the Storz Bickel Mighty medical vaporizer.  It an
opening for the screen & temperature control buttons and power button.
It has an allowance for the small triangular protrusion at the bottom of
the unit.

This does not allow for easy removal of the bottom tool nor access to
the power jack.