Account for nozzle distance
It was relative to the end of the heatsink, not the end of the nozzle,
which resulted in the bltouch being way too high.

The mount ended up in a tricky spot so the geometry around it is weird
to allow for the main mounting bracket's screws to have the space they
need (hopefully).